working front-end access for nutrient request
