+++ ORM/orm.js
@@ -1,0 +1,35 @@
+require("dotenv").config();
+const axios = require("axios");
+
+const getFood = (str) => {
+  return new Promise((resolve, reject) => {
+    axios({
+      method: "GET",
+      url: `https://api.edamam.com/api/food-database/v2/parser?app_id=${process.env.APP_ID}&app_key=${process.env.FOOD_API}&ingr=${str}`,
+    })
+      .then((response) => resolve(response))
+      .catch((err) => reject(err));
+  });
+};
+
+const getNutrients = (id, qty, unit) => {
+  return new Promise((resolve, reject) => {
+    axios({
+      method: "POST",
+      url: `https://api.edamam.com/api/food-database/v2/nutrients?app_id=${process.env.APP_ID}&app_key=${process.env.FOOD_API}`,
+      data: {
+        ingredients: [
+          {
+            quantity: parseInt(qty),
+            measureURI: `http://www.edamam.com/ontologies/edamam.owl#Measure_${unit}`,
+            foodId: id,
+          },
+        ],
+      },
+    })
+      .then((response) => resolve(response))
+      .catch((err) => reject(err));
+  });
+};
+
+module.exports = { getFood, getNutrients };
