+++ health-tracker/src/utils/API.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+export default {
+  getNutrients: (food, qty, unit) => {
+    return axios.get(`/food/${food}/${qty}/${unit}`);
+  },
+};
